% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfusionMatrix.R
\name{ConfusionMatrix}
\alias{ConfusionMatrix}
\title{ConfusionMatrix}
\usage{
ConfusionMatrix(beta, X, y)
}
\arguments{
\item{beta}{beta is a matrix}

\item{X}{X is a dataframe}

\item{y}{y is a dataframe}
}
\value{
a dataframe and other numeric scores
}
\description{
It can output the ConfusionMatrix of the model
}
\details{
This is some description of this function

You can use this function to output the ConfusionMatrix, accuracy, precision score, recall and F1 of the model.
}
\examples{
ConfusionMatrix(beta, X, y)


}
