% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costFunction.R
\name{costFunction}
\alias{costFunction}
\title{costFunction}
\usage{
costFunction(beta, X, y)
}
\arguments{
\item{beta}{beta is a matrix}

\item{X}{X is a dataframe, it is the attribute of the data set}

\item{y}{y is a dataframe, it is the label of the data set}
}
\value{
a number
}
\description{
It is used to calculate the costFunction of the logistic regression
}
\details{
This is some description of this function

 You can use this function to calculate costFunction, Jaccobi Matrix, and Hessian Matrix of the model
}
