% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GradientDescent.R
\name{GradientDescent}
\alias{GradientDescent}
\title{GradientDescent}
\usage{
GradientDescent(
  beta,
  X,
  y,
  maxLength = 10000,
  learningRate = 0.001,
  sigma = 0.001,
  regulation = 0,
  lamda = 0.5
)
}
\arguments{
\item{beta}{beta is a matrix}

\item{X}{X is a dataframe, it is the attribute of the data set}

\item{y}{y is a dataframe, it is the label of the data set}

\item{maxLength}{maxLength is a default param, which means the maximum length of the iteration.}

\item{learningRate}{learningRate is a number}

\item{sigma}{sigma is the threshold set to judge the stop condition.}

\item{regulation}{regulation is whether to add regularization to the model.}

\item{lamda}{lamda is the coefficient of the regulation term.}
}
\value{
a number
}
\description{
It is used to perform the GradientDescent model.
}
\details{
This is some description of this function

You can choose whether to add regularization to the model, by setting default regulation 0 to 2,
         which means you adopt, GradientDescent model with L2 norm regulation term.
}
