% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAUC.R
\name{plotAUC}
\alias{plotAUC}
\title{plotAUC}
\usage{
plotAUC(beta, X, y)
}
\arguments{
\item{beta}{beta is a matrix}

\item{X}{X is a dataframe, it is the attribute of the data set}

\item{y}{y is a dataframe, it is the label of the data set}
}
\value{
a number and plot ROC curve
}
\description{
It is a function used to plot the ROC curve of the model
}
\details{
This is some description of this function

You can use this function to plot the ROC curve and output the value of
         AUC by simply input the value of beta, X and y.
}
